requirements.txt
openai
anthropic
langgraph
langchain
langchain-openai

asyncpg
psycopg2-binary
sqlalchemy

pandas
numpy
python-dateutil

sqlparse

pinecone-client

fastapi
uvicorn[standard]
pydantic
python-multipart

redis
hiredis

python-dotenv
pyyaml
requests
faker